Add extension function to combine anti-air auras
